+++ http4k-cloudnative/src/test/kotlin/org/http4k/cloudnative/env/EnvironmentKeyTest.kt
@@ -7,13 +7,14 @@ import org.http4k.cloudnative.env.Environment.Companion.from
 import org.http4k.cloudnative.env.EnvironmentKey.k8s.HEALTH_PORT
 import org.http4k.cloudnative.env.EnvironmentKey.k8s.SERVICE_PORT
 import org.http4k.cloudnative.env.EnvironmentKey.k8s.serviceUriFor
+import org.http4k.cloudnative.env.EnvironmentKeyTest.Foo.bar
 import org.http4k.core.Uri
 import org.http4k.core.with
 import org.http4k.lens.LensFailure
 import org.http4k.lens.composite
 import org.http4k.lens.int
-import org.http4k.lens.of
 import org.http4k.lens.long
+import org.http4k.lens.of
 import org.junit.jupiter.api.Test
 import org.junit.jupiter.api.assertThrows
 import java.util.Properties
@@ -46,7 +47,10 @@ class EnvironmentKeyTest {
         assertThat(withInjectedValue["SOME_VALUE"], equalTo("80,81"))
 
         assertThat(EnvironmentKey.int().multi.required("SOME_VALUE")(withInjectedValue), equalTo(listOf(80, 81)))
-        assertThat(EnvironmentKey.int().multi.required("SOME_VALUE")(from("SOME_VALUE" to "80  , 81  ")), equalTo(listOf(80, 81)))
+        assertThat(
+            EnvironmentKey.int().multi.required("SOME_VALUE")(from("SOME_VALUE" to "80  , 81  ")),
+            equalTo(listOf(80, 81))
+        )
     }
 
     @Test
@@ -60,7 +64,13 @@ class EnvironmentKeyTest {
         assertThat(withInjectedValue["SOME_VALUE"], equalTo("80;81"))
 
         assertThat(EnvironmentKey.int().multi.required("SOME_VALUE")(withInjectedValue), equalTo(listOf(80, 81)))
-        assertThat(EnvironmentKey.int().multi.required("SOME_VALUE")(MapEnvironment.from(listOf("SOME_VALUE" to "80  ; 81  ").toMap().toProperties(), separator = ";")), equalTo(listOf(80, 81)))
+        assertThat(
+            EnvironmentKey.int().multi.required("SOME_VALUE")(
+                MapEnvironment.from(
+                    listOf("SOME_VALUE" to "80  ; 81  ").toMap().toProperties(), separator = ";"
+                )
+            ), equalTo(listOf(80, 81))
+        )
     }
 
     @Test
@@ -92,25 +102,40 @@ class EnvironmentKeyTest {
 
     @Test
     fun `get uri for a service`() {
-        assertThat(serviceUriFor("myservice")(
-            from("MYSERVICE_SERVICE_PORT" to "8000")),
-            equalTo(Uri.of("http://myservice:8000/")))
+        assertThat(
+            serviceUriFor("myservice")(
+                from("MYSERVICE_SERVICE_PORT" to "8000")
+            ),
+            equalTo(Uri.of("http://myservice:8000/"))
+        )
 
-        assertThat(serviceUriFor("myservice")(
-            from("MYSERVICE_SERVICE_PORT" to "80")),
-            equalTo(Uri.of("http://myservice/")))
+        assertThat(
+            serviceUriFor("myservice")(
+                from("MYSERVICE_SERVICE_PORT" to "80")
+            ),
+            equalTo(Uri.of("http://myservice/"))
+        )
 
-        assertThat(serviceUriFor("myservice", true)(
-            from("MYSERVICE_SERVICE_PORT" to "80")),
-            equalTo(Uri.of("https://myservice/")))
+        assertThat(
+            serviceUriFor("myservice", true)(
+                from("MYSERVICE_SERVICE_PORT" to "80")
+            ),
+            equalTo(Uri.of("https://myservice/"))
+        )
 
-        assertThat(serviceUriFor("myservice")(
-            from("MYSERVICE_SERVICE_PORT" to "443")),
-            equalTo(Uri.of("http://myservice/")))
+        assertThat(
+            serviceUriFor("myservice")(
+                from("MYSERVICE_SERVICE_PORT" to "443")
+            ),
+            equalTo(Uri.of("http://myservice/"))
+        )
 
-        assertThat(serviceUriFor("myservice", true)(
-            from("MYSERVICE_SERVICE_PORT" to "443")),
-            equalTo(Uri.of("https://myservice/")))
+        assertThat(
+            serviceUriFor("myservice", true)(
+                from("MYSERVICE_SERVICE_PORT" to "443")
+            ),
+            equalTo(Uri.of("https://myservice/"))
+        )
     }
 
     @Test
@@ -137,5 +162,16 @@ class EnvironmentKeyTest {
         }
 
         assertThat(key(finalEnv), equalTo(Target("bill", 123, null)))
+    }
+
+    enum class Foo {
+        bar
+    }
+
+    @Test
+    fun `enum support`() {
+        val key = EnvironmentKey.enum<Foo>().required("foo")
+        assertThat(key(EMPTY.with(key of bar)), equalTo(bar))
+
     }
 }